lessen the hardcore hearts effect
previously the hearts fully mimicked hardcore hearts. This feels a bit
disingenuous. Not wanting to "trick" anyone, we'll instead use a
lightened version of hardcore hearts, as this is really a "lesser"
version of hardcore.

Ideally, we'd use a completely different icon, but so far no appropriate
pixel-art has been created.